Assert action is introduced in IPerformanceTestCaseConfiguration interface
Test config now has an Assert action like it has a Prepare and Run action. The Assert action can be null (optional), but in case it is not null the Benchmarker will run it after the Run action. The Assert action got the config as parameter, similarly as Prepare and Run, so the Run can pass the result via the config instance to Assert. As Assert action is optional in the config this improvement is backward compatible, all existing performance test can remain intact.